requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
pandas==2.1.4
python-dotenv==1.0.0
Werkzeug==3.0.1
